+++ test/component_tests/text_input_test.exs
@@ -167,28 +167,6 @@ defmodule PrimerLive.TestComponents.TextInputTest do
     )
   end
 
-  test "Attribute: is_form_group (deprecatd)" do
-    assigns = %{}
-
-    run_test(
-      ~H"""
-      <.text_input form={:user} name="first_name" is_form_group />
-      """,
-      __ENV__
-    )
-  end
-
-  test "Attribute: is_form_group with input_id" do
-    assigns = %{}
-
-    run_test(
-      ~H"""
-      <.text_input form={:user} name="first_name" is_form_group input_id="xyz" />
-      """,
-      __ENV__
-    )
-  end
-
   test "Attribute: form_control (label)" do
     assigns = %{}
 
@@ -198,25 +176,6 @@ defmodule PrimerLive.TestComponents.TextInputTest do
         form={:user}
         field="first_name"
         form_control={
-          %{
-            label: "Some label"
-          }
-        }
-      />
-      """,
-      __ENV__
-    )
-  end
-
-  test "Attribute: form_group (label) (deprecated)" do
-    assigns = %{}
-
-    run_test(
-      ~H"""
-      <.text_input
-        form={:user}
-        field="first_name"
-        form_group={
           %{
             label: "Some label"
           }